iot2aps-lambda/iot2aps/requirements.txt
botocore
protobuf
python-snappy
requests
requests-aws4auth
